Fix animated shaders warping on high server time values
Remapped shaders and shader added with refEntity_t remain affected
(as long as they use timeOffset field). In original JK2 cgame there
are no shaders of second kind.

+++ src/renderer/tr_backend.cpp
@@ -543,7 +543,7 @@ void RB_RenderDrawSurfList( drawSurf_t *drawSurfs, int numDrawSurfs ) {
 	int				i;
 	drawSurf_t		*drawSurf;
 	unsigned int	oldSort;
-	float			originalTime;
+	double			originalTime;
 	bool			didShadowPass = false;
 
 	if (g_bRenderGlowingObjects)
@@ -743,7 +743,7 @@ void	RB_SetGL2D (void) {
 
 	// set time for 2D shaders
 	backEnd.refdef.time = ri.Milliseconds();
-	backEnd.refdef.floatTime = backEnd.refdef.time * 0.001f;
+	backEnd.refdef.floatTime = backEnd.refdef.time * 0.001;
 }
